Fix potential inline layouts color leakage.
Victor Porton reported this, with an fxnote stuck to a subsection with the
amsbook class. The subsection's title was colored as well. This seems to
occur when the subsection's text follows the title on the same line, and is
probably related to some parsing weirdness where a group would disappear.
I fixed this by half-blindly using \textcolor instead of \color directly.