Don't update the lox file in final mode anymore.
First, it's useless because the list of corrections is not typeset,
and second, this could lead to typesetting artifacts, even in final
mode which is totally unacceptable.

In short, contents line management is done by a non immediate \write
which inserts a whatsit in the vertical list. So even in final mode,
this whatsit could mess up with \addvspace.